add basic config infra

+++ tests/config_test.c
@@ -1,10 +1,10 @@
-#include <config.h>
+#include <config_interlal.h>
 #include <string.h>
 
 int main() {
     tmb_cfg_lexer_t lex = { 0 };
     const char* cfg     = "[loggers]\nhello = \"aaa%d%$\"";
-    tmb_lex(&lex, cfg, strlen(cfg));
+    tmb_lex(&lex, cfg, (int)strlen(cfg));
 
     for (int i = 0; i < lex.length; i++) { tmb_token_print(&lex.items[i]); }
     if (tmb_lex_expect(
